education institute site/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.4
email-validator==2.0.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
reportlab==4.0.4
Pillow==10.0.1
bcrypt==4.0.1
python-dateutil==2.8.2 